FDBD-47 fix[notification]:ddl-auto:update 환경에서 notification_metadata 데이터 타입 정합성 문제 해결

+++ src/main/java/com/mudosa/musinsa/notification/domain/dto/NoteMetaDTO.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class NoteMetaDTO {
-    private Long nMetadataId;
+    private Integer nMetadataId;
     private String notificationTitle;
     private String notificationMessage;
     private String notificationType;
